backend/requirements.txt
# FastAPI and web server
fastapi==0.100.*
gunicorn[uvicorn]==23.0.0
uvicorn==0.21.1
pydantic[email]==2.10.6
pydantic-settings==2.8.0
python-multipart==0.0.6
email-validator==2.0.0

# Database
sqlalchemy==2.0.7
alembic>=1.13.1,<1.14.0
psycopg2-binary>=2.9.9,<2.10.0

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Kubernetes
kubernetes==26.1.0

# Storage
boto3>=1.34.0,<1.35.0

# Notebook processing
nbformat>=5.9.2,<5.10.0
nbconvert>=7.14.0,<7.15.0
papermill>=2.5.0,<2.6.0
ipykernel>=6.29.0,<6.30.0

# Utilities
python-dateutil>=2.8.2,<2.9.0
tenacity>=8.2.3,<8.3.0
aiohttp>=3.9.1,<3.10.0
pyyaml>=6.0.1,<6.1.0
structlog>=23.1.0,<24.0.0
